requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
